% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_content.R
\name{cr_content}
\alias{cr_content}
\title{Conditionally rendering content}
\usage{
cr_content(content, only = NULL, cr2 = NULL)
}
\arguments{
\item{content}{content to be include conditionally. See \code{Details}}

\item{only}{see \code{\link{cr_val}}}

\item{cr2}{see \code{\link{cr_val}}}
}
\value{
markdwon content
}
\description{
Conditionally rendering content
}
\details{
Content can be anything you can insert in an common RMarkdown document,
except tha you can't insert \code{\`(backsticks)} and therefore chucks.
}
\examples{
\dontrun{
`r cr_text('Text to be evaluate.')`
}
}
