% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_cr2.R
\name{cr_cr2}
\alias{cr_cr2}
\title{Double conditional}
\usage{
cr_cr2(cr2)
}
\arguments{
\item{cr2}{character vector of format to be evaluate}
}
\value{
logical the opposite value of primary result for the given format
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
cr_cr2(cr2)
}
}
