% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_double.R
\name{cr_double}
\alias{cr_double}
\title{Double conditional}
\usage{
cr_double(cr2)
}
\arguments{
\item{cr2}{character vector of format to be evaluate}
}
\value{
logical the opposite value of primary result for the given format
}
\description{
Double conditional
}
\examples{
\dontrun{
cr_double(cr2)
}
}
