% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_format.R
\name{cr_format}
\alias{cr_format}
\title{Current knitting format}
\usage{
cr_format()
}
\value{
character Return de current knitting format. \code{html} if interactive.
}
\description{
\lifecycle{experimental}
}
