% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_header.R
\name{cr_header}
\alias{cr_header}
\title{Conditional headers}
\usage{
cr_header(htext, hlevel = 1, only = NULL, cr_dbl = NULL)
}
\arguments{
\item{htext}{character header text}

\item{hlevel}{numeric default to 1. TODO reference to markdown headers docs}

\item{only}{see \code{\link{cr_val}}}

\item{cr_dbl}{see \code{\link{cr_val}}}
}
\value{
markdown header conditionally
}
\description{
\lifecycle{soft-deprecated}
}
\examples{
\dontrun{
`r cr_header('Libraries')`
}
}
