% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_only.R
\name{cr_only}
\alias{cr_only}
\title{Only render on this formats}
\usage{
cr_only(only)
}
\arguments{
\item{only}{character a vector of character with the formats where the
content will be rendered.}
}
\value{
logical TRUE if current format in given vector, FALSE otherwise.
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
cr_only(only)
}
}
\keyword{internal}
