% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_text.R
\name{cr_text}
\alias{cr_text}
\title{Conditionally rendering text}
\usage{
cr_text(text, only = NULL, cr_dbl = NULL)
}
\arguments{
\item{text}{character text to be include conditionally}

\item{only}{see \code{\link{cr_val}}}

\item{cr_dbl}{see \code{\link{cr_val}}}
}
\value{
markdwon text
}
\description{
Conditionally rendering text
}
\examples{
\dontrun{
`r cr_text('Text to be evaluate.')`
}
}
