% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_val.R
\name{cr_val}
\alias{cr_val}
\title{Conditional value}
\usage{
cr_val(only = NULL, cr2 = NULL)
}
\arguments{
\item{only}{character default to \code{NULL}.
If specified, only in the specified formats the result will be \code{TRUE}}

\item{cr2}{character default to \code{NULL}.
If specified, the result for specified formats will be the opposite
to initial result.}
}
\value{
logical default to \code{TRUE} if format not in \code{yaml-params}.
Otherwise return a logical value based on given function params.
}
\description{
Conditional value
}
\examples{
\dontrun{
cr_val()
cr_val(only='html')
cr_val(cr_dbl='latex')
}
}
